+++ Johnson/ApiImplementation.cs
@@ -10,31 +10,13 @@ public sealed class ApiImplementation : IJohnsonApi
 	public IStatusEntry CrunchTimeStatus
 		=> ModEntry.Instance.CrunchTimeStatus;
 
-	public ICardTraitEntry TemporaryUpgradeCardTrait
-		=> TemporaryUpgradeManager.Trait;
-
-	public Tooltip TemporaryUpgradeTooltip
-		=> new GlossaryTooltip($"cardtrait.{ModEntry.Instance.Package.Manifest.UniqueName}::TemporaryUpgrade")
-		{
-			Icon = ModEntry.Instance.TemporaryUpgradeIcon.Sprite,
-			TitleColor = Colors.cardtrait,
-			Title = ModEntry.Instance.Localizations.Localize(["cardTrait", "TemporaryUpgrade", "name"]),
-			Description = ModEntry.Instance.Localizations.Localize(["cardTrait", "TemporaryUpgrade", "description"])
-		};
-
-	public bool IsTemporarilyUpgraded(Card card)
-		=> card.IsTemporarilyUpgraded();
-
-	public void SetTemporarilyUpgraded(Card card, bool value)
-		=> card.SetTemporarilyUpgraded(value);
-
 	public ICardTraitEntry StrengthenCardTrait
 		=> StrengthenManager.Trait;
 
 	public Tooltip GetStrengthenTooltip(int amount)
 		=> new GlossaryTooltip($"cardtrait.{ModEntry.Instance.Package.Manifest.UniqueName}::Strengthen")
 		{
-			Icon = ModEntry.Instance.TemporaryUpgradeIcon.Sprite,
+			Icon = ModEntry.Instance.StrengthenIcon.Sprite,
 			TitleColor = Colors.cardtrait,
 			Title = ModEntry.Instance.Localizations.Localize(["cardTrait", "Strengthen", "name"]),
 			Description = ModEntry.Instance.Localizations.Localize(["cardTrait", "Strengthen", "description"], new { Damage = amount })