+++ Johnson/Cards/MintCard.cs
@@ -65,11 +65,23 @@ internal sealed class MintCard : Card, IRegisterable
 						}
 					}
 				]
+			},
+			new ATooltipAction
+			{
+				Tooltips = upgrade switch
+				{
+					Upgrade.A => new UpgradeNonPermanentlyUpgradedCardBrowseAction().GetTooltips(s),
+					Upgrade.B => [new TTGlossary("cardtrait.temporary")],
+					_ => new MakeUpgradePermanentBrowseAction().GetTooltips(s),
+				}
 			}
 		];
 
 	public sealed class UpgradeNonPermanentlyUpgradedCardBrowseAction : CardAction
 	{
+		public override List<Tooltip> GetTooltips(State s)
+			=> [new TTGlossary("action.upgradeCard")];
+
 		public override Route? BeginWithRoute(G g, State s, Combat c)
 		{
 			timer = 0;
@@ -77,7 +89,7 @@ internal sealed class MintCard : Card, IRegisterable
 			if (selectedCard is null)
 				return baseResult;
 
-			selectedCard.SetTemporarilyUpgraded(false);
+			ModEntry.Instance.KokoroApi.TemporaryUpgrades.SetTemporaryUpgrade(selectedCard, null);
 			return new CardUpgrade
 			{
 				cardCopy = Mutil.DeepCopy(selectedCard)
@@ -87,15 +99,22 @@ internal sealed class MintCard : Card, IRegisterable
 
 	public sealed class MakeUpgradePermanentBrowseAction : CardAction
 	{
+		public override List<Tooltip> GetTooltips(State s)
+			=> [
+				new TTGlossary("action.upgradeCard"),
+				ModEntry.Instance.KokoroApi.TemporaryUpgrades.UpgradeTooltip,
+			];
+
 		public override void Begin(G g, State s, Combat c)
 		{
 			base.Begin(g, s, c);
 			timer = 0;
 			if (selectedCard is null)
 				return;
-			if (selectedCard.upgrade == Upgrade.None || !selectedCard.IsTemporarilyUpgraded())
+			if (ModEntry.Instance.KokoroApi.TemporaryUpgrades.GetTemporaryUpgrade(selectedCard) is not { } temporaryUpgrade)
 				return;
-			selectedCard.SetTemporarilyUpgraded(false);
+			ModEntry.Instance.KokoroApi.TemporaryUpgrades.SetTemporaryUpgrade(selectedCard, null);
+			ModEntry.Instance.KokoroApi.TemporaryUpgrades.SetPermanentUpgrade(selectedCard, temporaryUpgrade);
 		}
 	}
 }