+++ Johnson/ExternalAPIs/Kokoro/Actions.cs
@@ -6,7 +6,7 @@ public partial interface IKokoroApi
 {
 	IActionApi Actions { get; }
 
-	public interface IActionApi
+	public partial interface IActionApi
 	{
 		ACardOffering WithDestination(ACardOffering action, CardDestination? destination, bool? insertRandomly = null);
 		CardReward WithDestination(CardReward route, CardDestination? destination, bool? insertRandomly = null);