+++ Johnson/IJohnsonApi.cs
@@ -6,11 +6,6 @@ public interface IJohnsonApi
 {
 	IDeckEntry JohnsonDeck { get; }
 	IStatusEntry CrunchTimeStatus { get; }
-	
-	ICardTraitEntry TemporaryUpgradeCardTrait { get; }
-	Tooltip TemporaryUpgradeTooltip { get; }
-	bool IsTemporarilyUpgraded(Card card);
-	void SetTemporarilyUpgraded(Card card, bool value);
 
 	ICardTraitEntry StrengthenCardTrait { get; }
 	Tooltip GetStrengthenTooltip(int amount);