+++ Johnson/ModEntry.cs
@@ -34,7 +34,6 @@ public sealed class ModEntry : SimpleMod
 	internal IPlayableCharacterEntryV2 JohnsonCharacter { get; }
 	internal IStatusEntry CrunchTimeStatus { get; }
 
-	internal ISpriteEntry TemporaryUpgradeIcon { get; }
 	internal ISpriteEntry StrengthenIcon { get; }
 	internal ISpriteEntry StrengthenHandIcon { get; }
 	internal ISpriteEntry DiscountHandIcon { get; }
@@ -122,7 +121,7 @@ public sealed class ModEntry : SimpleMod
 	public ModEntry(IPluginPackage<IModManifest> package, IModHelper helper, ILogger logger) : base(package, helper, logger)
 	{
 		Instance = this;
-		Harmony = helper.Utilities.DelayedHarmony;
+		Harmony = helper.Utilities.Harmony;
 		KokoroApi = helper.ModRegistry.GetApi<IKokoroApi>("Shockah.Kokoro")!;
 		DuoArtifactsApi = helper.ModRegistry.GetApi<IDuoArtifactsApi>("Shockah.DuoArtifacts");
 
@@ -147,11 +146,9 @@ public sealed class ModEntry : SimpleMod
 
 		_ = new CrunchTimeManager();
 		_ = new StrengthenManager();
-		_ = new TemporaryUpgradeManager();
 
 		DynamicWidthCardAction.ApplyPatches(Harmony, logger);
 		CustomCardBrowse.ApplyPatches(Harmony, logger);
-		InPlaceCardUpgrade.ApplyPatches(Harmony, logger);
 
 		CustomCardBrowse.RegisterCustomCardSource(
 			UpgradableCardsInHandBrowseSource,
@@ -171,7 +168,7 @@ public sealed class ModEntry : SimpleMod
 			TemporarilyUpgradedCardsBrowseSource,
 			new CustomCardBrowse.CustomCardSource(
 				(_, _, _) => Loc.T("cardBrowse.title.upgrade"),
-				(state, combat) => state.GetAllCards().Where(c => c.upgrade != Upgrade.None && c.IsTemporarilyUpgraded()).ToList()
+				(state, _) => state.GetAllCards().Where(c => c.upgrade != Upgrade.None && KokoroApi.TemporaryUpgrades.GetTemporaryUpgrade(c) is not null).ToList()
 			)
 		);
 		CustomCardBrowse.RegisterCustomCardSource(
@@ -199,7 +196,7 @@ public sealed class ModEntry : SimpleMod
 			NonPermanentlyUpgradedCardsBrowseSource,
 			new CustomCardBrowse.CustomCardSource(
 				(_, _, _) => Loc.T("cardBrowse.title.upgrade"),
-				(state, combat) => state.deck.Concat(combat?.discard ?? []).Concat(combat?.hand ?? []).Where(c => c.upgrade == Upgrade.None || c.IsTemporarilyUpgraded()).ToList()
+				(state, combat) => state.deck.Concat(combat?.discard ?? []).Concat(combat?.hand ?? []).Where(c => KokoroApi.TemporaryUpgrades.GetPermanentUpgrade(c) == Upgrade.None).ToList()
 			)
 		);
 		CustomCardBrowse.RegisterCustomCardSource(
@@ -305,7 +302,6 @@ public sealed class ModEntry : SimpleMod
 				.ToList()
 		});
 
-		TemporaryUpgradeIcon = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Icons/TemporaryUpgrade.png"));
 		StrengthenIcon = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Icons/Strengthen.png"));
 		StrengthenHandIcon = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Icons/StrengthenHand.png"));
 		DiscountHandIcon = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Icons/DiscountHand.png"));