+++ Kokoro/API/Actions/CustomCardUpgrade.cs
@@ -1,0 +1,18 @@
+﻿namespace Shockah.Kokoro;
+
+public partial interface IKokoroApi
+{
+	public partial interface IActionApi
+	{
+		ICustomCardUpgrade MakeCustomCardUpgrade(CardUpgrade route);
+
+		public interface ICustomCardUpgrade
+		{
+			CardUpgrade AsRoute { get; }
+			
+			bool IsInPlace { get; set; }
+
+			ICustomCardUpgrade SetInPlace(bool value);
+		}
+	}
+}